git-debimport openmpi_1.6.5-8.debian.tar.gz

+++ ompi/mca/btl/tcp/btl_tcp_frag.c
@@ -41,7 +41,6 @@
 #include "ompi/mca/btl/base/btl_base_error.h"
 #include "btl_tcp_frag.h" 
 #include "btl_tcp_endpoint.h"
-#include "orte/util/proc_info.h"
 
 static void mca_btl_tcp_frag_common_constructor(mca_btl_tcp_frag_t* frag) 
 { 
@@ -113,9 +112,9 @@ bool mca_btl_tcp_frag_send(mca_btl_tcp_frag_t* frag, int sd)
             case EWOULDBLOCK:
                 return false;
             case EFAULT:
-                BTL_ERROR(("mca_btl_tcp_frag_send: writev error (%p, %d)\n\t%s(%d)\n",
-                    frag->iov_ptr[0].iov_base, frag->iov_ptr[0].iov_len,
-                    strerror(opal_socket_errno), frag->iov_cnt));
+                BTL_ERROR(("mca_btl_tcp_frag_send: writev error (%p, %lu)\n\t%s(%lu)\n",
+                    frag->iov_ptr[0].iov_base, (unsigned long) frag->iov_ptr[0].iov_len,
+                    strerror(opal_socket_errno), (unsigned long) frag->iov_cnt));
                 mca_btl_tcp_endpoint_close(frag->endpoint);
                 return false;
             default:
@@ -205,9 +204,9 @@ bool mca_btl_tcp_frag_recv(mca_btl_tcp_frag_t* frag, int sd)
 	case EWOULDBLOCK:
 	    return false;
 	case EFAULT:
-            BTL_ERROR(("mca_btl_tcp_frag_recv: readv error (%p, %d)\n\t%s(%d)\n",
-                       frag->iov_ptr[0].iov_base, frag->iov_ptr[0].iov_len,
-                       strerror(opal_socket_errno), frag->iov_cnt));
+            BTL_ERROR(("mca_btl_tcp_frag_recv: readv error (%p, %lu)\n\t%s(%lu)\n",
+                       frag->iov_ptr[0].iov_base, (unsigned long) frag->iov_ptr[0].iov_len,
+                       strerror(opal_socket_errno), (unsigned long) frag->iov_cnt));
 	    mca_btl_tcp_endpoint_close(btl_endpoint);
 	    return false;
 	default:
